重新配置了数据库 改用了阿里数据库 弃用服务器搭建的数据库 部门自己联查自己

+++ tabac/src/main/java/com/xr/controller/SysUserController.java
@@ -135,7 +135,7 @@ public class SysUserController {
 
     @RequestMapping("check")
     public ResponseResult check(String username){
-        int number=sysUserService.checkname(username);
+        int number= Math.toIntExact(sysUserService.checkname(username));
         ResponseResult result=new ResponseResult();
         result.getData().put("number",number);
         return result;
